Move event code to new *HookEvent classes
