+++ src/main/java/hudson/plugins/tfs/TeamWebHook.java
@@ -151,111 +151,6 @@ public class TeamWebHook implements UnprotectedRootAction {
         return dispatch(request, body);
     }
 
-    public List<GitStatus.ResponseContributor> pullRequestMergeCommitCreated(final PullRequestMergeCommitCreatedEventArgs args) {
-        final PullRequestParameterAction action = new PullRequestParameterAction(args);
-        // TODO: add extension point for this event, then extract current implementation as extension(s)
-
-        return pollOrQueueFromEvent(args, action);
-    }
-
-    public List<GitStatus.ResponseContributor> gitCodePushed(final GitCodePushedEventArgs gitCodePushedEventArgs) {
-        final CommitParameterAction commitParameterAction = new CommitParameterAction(gitCodePushedEventArgs);
-        // TODO: add extension point for this event, then extract current implementation as extension(s)
-
-        return pollOrQueueFromEvent(gitCodePushedEventArgs, commitParameterAction);
-    }
-
-    List<GitStatus.ResponseContributor> pollOrQueueFromEvent(final GitCodePushedEventArgs gitCodePushedEventArgs, final CommitParameterAction commitParameterAction) {
-        List<GitStatus.ResponseContributor> result = new ArrayList<GitStatus.ResponseContributor>();
-        final String commit = gitCodePushedEventArgs.commit;
-        final URIish uri = gitCodePushedEventArgs.getRepoURIish();
-
-        // run in high privilege to see all the projects anonymous users don't see.
-        // this is safe because when we actually schedule a build, it's a build that can
-        // happen at some random time anyway.
-        SecurityContext old = ACL.impersonate(ACL.SYSTEM);
-        try {
-
-            boolean scmFound = false;
-            final Jenkins jenkins = Jenkins.getInstance();
-            if (jenkins == null) {
-                LOGGER.severe("Jenkins.getInstance() is null");
-                return result;
-            }
-            for (final Item project : Jenkins.getInstance().getAllItems()) {
-                final SCMTriggerItem scmTriggerItem = SCMTriggerItem.SCMTriggerItems.asSCMTriggerItem(project);
-                if (scmTriggerItem == null) {
-                    continue;
-                }
-                for (final SCM scm : scmTriggerItem.getSCMs()) {
-                    if (!(scm instanceof GitSCM)) {
-                        continue;
-                    }
-                    final GitSCM git = (GitSCM) scm;
-                    scmFound = true;
-
-                    for (final RemoteConfig repository : git.getRepositories()) {
-                        boolean repositoryMatches = false;
-                        for (URIish remoteURL : repository.getURIs()) {
-                            if (GitStatus.looselyMatches(uri, remoteURL)) {
-                                repositoryMatches = true;
-                                break;
-                            }
-                        }
-
-                        if (!repositoryMatches || git.getExtensions().get(IgnoreNotifyCommit.class)!=null) {
-                            continue;
-                        }
-
-                        if (!(project instanceof AbstractProject && ((AbstractProject) project).isDisabled())) {
-                            if (project instanceof Job) {
-                                // TODO: Add default parameters defined in the job
-                                final Job job = (Job) project;
-
-                                boolean triggered = false;
-                                if (!triggered) {
-                                    // TODO: check global override here
-                                }
-
-                                if (!triggered) {
-                                    final SCMTrigger scmTrigger = findTrigger(job, SCMTrigger.class);
-                                    if (scmTrigger != null && !scmTrigger.isIgnorePostCommitHooks()) {
-                                        // queue build without first polling
-                                        final int quietPeriod = scmTriggerItem.getQuietPeriod();
-                                        final Cause cause = new TeamHookCause(commit);
-                                        final CauseAction causeAction = new CauseAction(cause);
-                                        scmTriggerItem.scheduleBuild2(quietPeriod, causeAction, commitParameterAction);
-                                        result.add(new ScheduledResponseContributor(project));
-                                        triggered = true;
-                                    }
-                                }
-                                if (!triggered) {
-                                    final TeamPushTrigger pushTrigger = findTrigger(job, TeamPushTrigger.class);
-                                    if (pushTrigger != null) {
-                                        pushTrigger.execute(gitCodePushedEventArgs, commitParameterAction);
-                                        result.add(new PollingScheduledResponseContributor(project));
-                                        triggered = true;
-                                    }
-                                }
-                                if (triggered) {
-                                    break;
-                                }
-                            }
-                        }
-                    }
-                }
-            }
-            if (!scmFound) {
-                result.add(new GitStatus.MessageResponseContributor("No Git jobs found"));
-            }
-
-            return result;
-        }
-        finally {
-            SecurityContextHolder.setContext(old);
-        }
-    }
-
     public static <T extends Trigger> T findTrigger(final Job<?, ?> job, final Class<T> tClass) {
         if (job instanceof ParameterizedJobMixIn.ParameterizedJob) {
             final ParameterizedJobMixIn.ParameterizedJob pJob = (ParameterizedJobMixIn.ParameterizedJob) job;
